+++ .eslintrc
@@ -1,4 +1,4 @@
 {
   "extends": "airbnb-base",
-  "rules": {"no-underscore-dangle": "off"}
+  "rules": {"no-underscore-dangle": ["error", { "allow": ["_id"]}]}
 }
