+++ README.md
@@ -1,2 +1,10 @@
 # movies-explorer-api
 graduation project (backend)
+
+------------------------------------------------
+|'http://kino-domino.nomoredomains.rocks'      |
+------------------------------------------------   
+|'https://kino-domino.nomoredomains.rocks'     |
+------------------------------------------------
+|51.250.3.195                                  |
+------------------------------------------------
